refactor: separate source files into dedicated package to structure source code
Application.kt - Entry point
- model/
    - Car.kt
- view/
    - InputView.kt
    - ResultView.kt
- controller/
    - RacingManager.kt
- util
    - RandomValueGenerator.kt
