+++ src/main/kotlin/racingcar/util/RandomValueGenerator.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.util
 
 import camp.nextstep.edu.missionutils.Randoms
 
@@ -6,14 +6,14 @@ interface ValueGenerator {
     fun generateNumber(): Int
 }
 
-object RandomValueGenerator : ValueGenerator{
+object RandomValueGenerator : ValueGenerator {
     override fun generateNumber(): Int = Randoms.pickNumberInRange(0, 9)
 }
 
-object TestMoveValueGenerator : ValueGenerator{
+object TestMoveValueGenerator : ValueGenerator {
     override fun generateNumber(): Int = 9
 }
 
-object TestDoNotMoveValueGenerator : ValueGenerator{
+object TestDoNotMoveValueGenerator : ValueGenerator {
     override fun generateNumber(): Int = 0
 }
