+++ src/test/kotlin/racingcar/RacingManagerTest.kt
@@ -2,6 +2,10 @@ package racingcar
 
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.Assertions.assertEquals
+import racingcar.controller.RacingManager
+import racingcar.model.Car
+import racingcar.util.TestDoNotMoveValueGenerator
+import racingcar.util.TestMoveValueGenerator
 
 class RacingManagerTest {
     @Test
