+++ src/test/kotlin/racingcar/RandomValueGeneratorTest.kt
@@ -2,6 +2,7 @@ package racingcar
 
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.Assertions.assertEquals
+import racingcar.util.RandomValueGenerator
 
 class RandomValueGeneratorTest {
     @Test
